Move test_evalcoeffs to separate files
